Corrected bug where if the last coin used to purchase a soda was a penny the change would be an outrageously large decimal number by adding round to a certain decimal number method.

+++ SodaMachine/Penny.cs
@@ -14,7 +14,7 @@ namespace SodaMachine
         public Penny()
         {
             Name = "penny";
-            value = 0.01;
+            value = (double)(decimal)0.01;
         }
         //Member Methods (Can Do)
     }
